# TAS-20: Ruby FFI Optimization with Magnus Wrapped Classes