… CLAUDE.md
… bindings/ruby/lib/tasker_core/orchestration_manager.rb
… bindings/ruby/lib/tasker_core/task_handler.rb
… bindings/ruby/spec/examples.txt
… bindings/ruby/test_comprehensive_ffi_optimization.rb
… bindings/ruby/test_optimized_ffi.rb
… bindings/ruby/test_thread_safety.rb
… bindings/ruby/test_thread_safety_comparison.rb
… docs/roadmap/README.md